+++ server/src/main/java/org/opensearch/index/fielddata/FieldDataStats.java
@@ -32,6 +32,8 @@
 
 package org.opensearch.index.fielddata;
 
+import org.opensearch.Version;
+import org.opensearch.common.FieldCountStats;
 import org.opensearch.common.FieldMemoryStats;
 import org.opensearch.common.Nullable;
 import org.opensearch.common.annotation.PublicApi;
@@ -43,7 +45,9 @@ import org.opensearch.core.xcontent.ToXContentFragment;
 import org.opensearch.core.xcontent.XContentBuilder;
 
 import java.io.IOException;
+import java.util.List;
 import java.util.Objects;
+import java.util.Set;
 
 /**
  * Encapsulates heap usage for field data
@@ -58,10 +62,17 @@ public class FieldDataStats implements Writeable, ToXContentFragment {
     private static final String MEMORY_SIZE_IN_BYTES = "memory_size_in_bytes";
     private static final String EVICTIONS = "evictions";
     private static final String FIELDS = "fields";
+    private static final String ITEM_COUNT = "item_count";
     private long memorySize;
     private long evictions;
+    private long itemCount;
     @Nullable
-    private FieldMemoryStats fields;
+    private FieldMemoryStats fieldMemorySizes;
+    @Nullable
+    private FieldCountStats fieldItemCounts;
+
+    private static final List<String> ORDERED_FIELD_LEVEL_API_RAW_KEYS = List.of(MEMORY_SIZE_IN_BYTES, ITEM_COUNT);
+    private static final List<String> ORDERED_FIELD_LEVEL_API_READABLE_KEYS = List.of(MEMORY_SIZE, ITEM_COUNT);
 
     public FieldDataStats() {
 
@@ -70,13 +81,39 @@ public class FieldDataStats implements Writeable, ToXContentFragment {
     public FieldDataStats(StreamInput in) throws IOException {
         memorySize = in.readVLong();
         evictions = in.readVLong();
-        fields = in.readOptionalWriteable(FieldMemoryStats::new);
+        fieldMemorySizes = in.readOptionalWriteable(FieldMemoryStats::new);
+        if (in.getVersion().onOrAfter(Version.V_3_4_0)) {
+            itemCount = in.readVLong();
+            fieldItemCounts = in.readOptionalWriteable(FieldCountStats::new);
+        }
     }
 
+    /**
+     * @deprecated Use the constructor that includes item counts instead.
+     * @param memorySize the total memory size
+     * @param evictions the number of evictions
+     * @param fields the memory size for each field
+     */
+    @Deprecated
     public FieldDataStats(long memorySize, long evictions, @Nullable FieldMemoryStats fields) {
+        this(memorySize, evictions, fields, 0, null);
+    }
+
+    public FieldDataStats(
+        long memorySize,
+        long evictions,
+        @Nullable FieldMemoryStats fieldMemorySizes,
+        long itemCount,
+        @Nullable FieldCountStats fieldItemCounts
+    ) {
         this.memorySize = memorySize;
         this.evictions = evictions;
-        this.fields = fields;
+        this.fieldMemorySizes = fieldMemorySizes;
+        this.itemCount = itemCount;
+        this.fieldItemCounts = fieldItemCounts;
+        if (fieldMemorySizes != null && fieldItemCounts != null) {
+            assert fieldMemorySizes.getStats().keySet().equals(fieldItemCounts.getStats().keySet());
+        }
     }
 
     public void add(FieldDataStats stats) {
@@ -85,11 +122,19 @@ public class FieldDataStats implements Writeable, ToXContentFragment {
         }
         this.memorySize += stats.memorySize;
         this.evictions += stats.evictions;
-        if (stats.fields != null) {
-            if (fields == null) {
-                fields = stats.fields.copy();
+        this.itemCount += stats.itemCount;
+        if (stats.fieldMemorySizes != null) {
+            if (fieldMemorySizes == null) {
+                fieldMemorySizes = stats.fieldMemorySizes.copy();
             } else {
-                fields.add(stats.fields);
+                fieldMemorySizes.add(stats.fieldMemorySizes);
+            }
+        }
+        if (stats.fieldItemCounts != null) {
+            if (fieldItemCounts == null) {
+                fieldItemCounts = stats.fieldItemCounts.copy();
+            } else {
+                fieldItemCounts.add(stats.fieldItemCounts);
             }
         }
     }
@@ -106,16 +151,34 @@ public class FieldDataStats implements Writeable, ToXContentFragment {
         return this.evictions;
     }
 
+    public long getItemCount() {
+        return this.itemCount;
+    }
+
+    /**
+     * Get the field memory sizes.
+     * Cannot change the method name as this class was marked PublicApi.
+     * @return The memory field sizes
+     */
     @Nullable
     public FieldMemoryStats getFields() {
-        return fields;
+        return fieldMemorySizes;
+    }
+
+    @Nullable
+    public FieldMemoryStats getFieldItemCounts() {
+        return fieldItemCounts;
     }
 
     @Override
     public void writeTo(StreamOutput out) throws IOException {
         out.writeVLong(memorySize);
         out.writeVLong(evictions);
-        out.writeOptionalWriteable(fields);
+        out.writeOptionalWriteable(fieldMemorySizes);
+        if (out.getVersion().onOrAfter(Version.V_3_4_0)) {
+            out.writeVLong(itemCount);
+            out.writeOptionalWriteable(fieldItemCounts);
+        }
     }
 
     @Override
@@ -123,8 +186,23 @@ public class FieldDataStats implements Writeable, ToXContentFragment {
         builder.startObject(FIELDDATA);
         builder.humanReadableField(MEMORY_SIZE_IN_BYTES, MEMORY_SIZE, getMemorySize());
         builder.field(EVICTIONS, getEvictions());
+        builder.field(ITEM_COUNT, getItemCount());
+        Set<String> fields = null;
+        // We've already asserted the two key sets are the same if both are not null
+        if (fieldMemorySizes != null) {
+            fields = fieldMemorySizes.getStats().keySet();
+        } else if (fieldItemCounts != null) {
+            fields = fieldItemCounts.getStats().keySet();
+        }
         if (fields != null) {
-            fields.toXContent(builder, FIELDS, MEMORY_SIZE_IN_BYTES, MEMORY_SIZE);
+            FieldMemoryStats.toXContent(
+                builder,
+                fields,
+                FIELDS,
+                List.of(fieldMemorySizes, fieldItemCounts),
+                ORDERED_FIELD_LEVEL_API_RAW_KEYS,
+                ORDERED_FIELD_LEVEL_API_READABLE_KEYS
+            );
         }
         builder.endObject();
         return builder;
@@ -135,11 +213,15 @@ public class FieldDataStats implements Writeable, ToXContentFragment {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
         FieldDataStats that = (FieldDataStats) o;
-        return memorySize == that.memorySize && evictions == that.evictions && Objects.equals(fields, that.fields);
+        return memorySize == that.memorySize
+            && evictions == that.evictions
+            && Objects.equals(fieldMemorySizes, that.fieldMemorySizes)
+            && itemCount == that.itemCount
+            && Objects.equals(fieldItemCounts, that.fieldItemCounts);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(memorySize, evictions, fields);
+        return Objects.hash(memorySize, evictions, fieldMemorySizes, itemCount, fieldItemCounts);
     }
 }